pyproject.toml
[project]
name = "linux-ssh-mcp"
version = "0.1.0"
description = "Linux SSH MCP远程命令执行工具"
requires-python = ">=3.10"
dependencies = [
  "asyncssh>=2.13.0",
  "keyring>=24.0.0",
  "loguru>=0.7.0",
  "mcp>=1.0.0",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.8.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.3",
  "pytest-mock>=3.12.0",
  "ruff>=0.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
files = ["linux_ssh_mcp", "main.py", "diagnose_mcp.py", "trae_mcp_bootstrap.py"]
exclude = "^(tests[\\/]|create_ide_config\\.py$|debug_server\\.py$|diagnose_startup\\.py$|direct_test\\.py$|file_test\\.py$|main_fixed\\.py$|minimal_server\\.py$|minimal_test\\.py$|quick_test\\.py$|start_server\\.py$|test_mcp_protocol\\.py$|utf8_test\\.py$)"

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [
  "create_ide_config.py",
  "debug_server.py",
  "diagnose_startup.py",
  "direct_test.py",
  "file_test.py",
  "main_fixed.py",
  "minimal_server.py",
  "minimal_test.py",
  "quick_test.py",
  "start_server.py",
  "test_mcp_protocol.py",
  "utf8_test.py",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
